Main turn/battle logic
In this i made some basic switch cases for the menu so now player 1 and player 2 can battle it out. I could have maybe made a class for the players and handled the logic in there to eliminate repetition  of the code in the main function. but it works! lol