Small changes and win condition
For some reason my first commit isn't here as i had a merge issue? anyway it just showed the basic functionality without player turns.

In this commit i added a win condition as i forgot that before as well as a couple new messages for the player to read better